The rime channel header size is now stored in uint16_t.
If using an 8 byte rime address size in the multihop example,
an 261 bit header is created. This produced an overflow.

+++ core/net/rime/channel.h
@@ -50,7 +50,7 @@ struct channel {
   struct channel *next;
   uint16_t channelno;
   const struct packetbuf_attrlist *attrlist;
-  uint8_t hdrsize;
+  uint16_t hdrsize;
 };
 
 struct channel *channel_lookup(uint16_t channelno);
